@@ -5,7 +5,7 @@
   "main": "host/electron.js",
   "build": {
     "appId": "com.microsoft.azure.iot.pnp.ui",
-    "productName": "Azure IoT Explorer (preview)",
+    "productName": "Azure IoT Explorer Preview",
     "files": [
       "dist/**/*",
       "package.json",
@@ -76,6 +76,7 @@
     "body-parser": "1.18.3",
     "brace": "0.11.1",
     "cors": "2.8.5",
+    "crypto-browserify": "^3.12.0",
     "date-fns": "2.14.0",
     "electron-window-state": "5.0.3",
     "express": "4.16.4",
@@ -110,7 +111,7 @@
     "@types/i18next": "8.4.4",
     "@types/jest": "24.0.15",
     "@types/jest-plugin-context": "2.9.0",
-    "@types/node": "14.0.27",
+    "@types/node": "16.11.26",
     "@types/react": "16.9.35",
     "@types/react-dom": "16.9.8",
     "@types/react-jsonschema-form": "1.0.10",
@@ -118,45 +119,42 @@
     "@types/request": "2.48.1",
     "@types/semver": "6.0.2",
     "@types/uuid": "3.4.5",
-    "@types/webpack": "4.41.27",
-    "@types/webpack-dev-server": "3.11.3",
-    "@types/webpack-merge": "4.1.5",
-    "awesome-typescript-loader": "5.2.1",
     "concurrently": "5.3.0",
-    "copy-webpack-plugin": "6.1.0",
-    "css-loader": "1.0.0",
-    "electron": "11.5.0",
-    "electron-builder": "22.8.0",
+    "copy-webpack-plugin": "6.4.1",
+    "css-loader": "6.7.1",
+    "css-minimizer-webpack-plugin": "3.4.1",
+    "electron": "18.0.0",
+    "electron-builder": "22.14.13",
     "enzyme": "3.11.0",
     "enzyme-adapter-react-16": "1.15.1",
     "enzyme-to-json": "3.3.5",
     "file-loader": "2.0.0",
-    "html-webpack-plugin": "3.2.0",
-    "is-svg": "^4.3.1",
+    "html-webpack-plugin": "5.5.0",
+    "is-svg": "4.3.2",
     "jest": "26.4.2",
     "jest-plugin-context": "2.9.0",
     "jest-trx-results-processor": "0.0.7",
     "lodash": "4.17.21",
-    "mini-css-extract-plugin": "0.8.0",
+    "mini-css-extract-plugin": "2.6.0",
+    "node-polyfill-webpack-plugin": "1.1.4",
     "nodemon": "2.0.4",
-    "optimize-css-assets-webpack-plugin": "5.0.3",
     "sass": "1.35.2",
     "sass-loader": "7.2.0",
     "source-map-loader": "0.2.4",
-    "ssri": "^8.0.1",
+    "ssri": "8.0.1",
     "style-loader": "0.23.1",
-    "terser-webpack-plugin": "1.4.5",
     "ts-jest": "26.2.0",
+    "ts-loader": "9.2.8",
     "tslint": "5.11.0",
     "tslint-loader": "3.5.4",
     "tslint-origin-ordered-imports-rule": "1.1.2",
     "tslint-react": "3.6.0",
-    "typescript": "3.8.3",
-    "webpack": "4.41.6",
+    "typescript": "4.6.3",
+    "webpack": "5.70.0",
     "webpack-bundle-analyzer": "3.3.2",
-    "webpack-cli": "3.3.12",
-    "webpack-dev-server": "3.11.0",
-    "webpack-merge": "4.2.1",
+    "webpack-cli": "4.9.2",
+    "webpack-dev-server": "4.7.4",
+    "webpack-merge": "5.8.0",
     "webpack-shell-plugin": "0.5.0"
   },
   "jest": {
