@@ -3,6 +3,7 @@
  * Licensed under the MIT License
  **********************************************************/
 import { Action } from 'typescript-fsa';
+import { SagaIterator } from 'redux-saga';
 import { call, put } from 'redux-saga/effects';
 import { raiseNotificationToast } from '../../../notifications/components/notificationToast';
 import { NotificationType } from '../../../api/models/notification';
@@ -14,7 +15,7 @@ import { ERROR_TYPES } from './../../../constants/apiConstants';
 import { appConfig } from '../../../../appConfig/appConfig';
 import { CUSTOM_CONTROLLER_PORT } from './../../../constants/browserStorage';
 
-export function* listDevicesSaga(action: Action<DeviceQuery>) {
+export function* listDevicesSaga(action: Action<DeviceQuery>): SagaIterator {
     try {
         const parameters = {
             query: action.payload
