@@ -3,6 +3,7 @@
  * Licensed under the MIT License
  **********************************************************/
 import { call, put, takeLatest } from 'redux-saga/effects';
+import { SagaIterator } from 'redux-saga';
 import { Action } from 'typescript-fsa';
 import { fetchModuleIdentities } from '../../../api/services/moduleService';
 import { raiseNotificationToast } from '../../../notifications/components/notificationToast';
@@ -10,7 +11,7 @@ import { NotificationType } from '../../../api/models/notification';
 import { ResourceKeys } from '../../../../localization/resourceKeys';
 import { getModuleIdentitiesAction } from './actions';
 
-export function* getModuleIdentitiesSagaWorker(action: Action<string>) {
+export function* getModuleIdentitiesSagaWorker(action: Action<string>): SagaIterator {
     try {
         const parameters = {
             deviceId: action.payload
