@@ -3,6 +3,7 @@
  * Licensed under the MIT License
  **********************************************************/
 import { call } from 'redux-saga/effects';
+import { SagaIterator } from 'redux-saga';
 import { Action } from 'typescript-fsa';
 import { setConfigurableRepositoryPath, setLocalFolderPath, setRepositoryLocations } from '../../../api/services/modelRepositoryService';
 import { RepositoryLocationSettings } from '../state';
@@ -11,7 +12,7 @@ import { NotificationType } from '../../../api/models/notification';
 import { ResourceKeys } from '../../../../localization/resourceKeys';
 import { raiseNotificationToast } from '../../../notifications/components/notificationToast';
 
-export function* setRepositoryLocationsSaga(action: Action<RepositoryLocationSettings[]>) {
+export function* setRepositoryLocationsSaga(action: Action<RepositoryLocationSettings[]>): SagaIterator {
     const localFolderPath = yield call(getLocalFolderPath, action.payload);
     yield call(setLocalFolderPath, localFolderPath);
 
